ProjectileManager - same as TileManager but handles Projectiles
